- Started working on OrderMananger. - Added RemoteOrderService and it's implementation. - Added multiple DTOs. - refactored packages. - added Converters, so multiple manager can use them.

+++ fbs-remote/src/main/java/ch/hslu/appe/fbs/remote/utils/ArticlePriceDescComparator.java
@@ -1,6 +1,6 @@
 package ch.hslu.appe.fbs.remote.utils;
 
-import ch.hslu.appe.fbs.remote.ArticleDTO;
+import ch.hslu.appe.fbs.remote.dtos.ArticleDTO;
 
 import java.util.Comparator;
 
@@ -12,6 +12,6 @@ import java.util.Comparator;
 public final class ArticlePriceDescComparator implements Comparator<ArticleDTO> {
     @Override
     public int compare(final ArticleDTO o1, final ArticleDTO o2) {
-        return Float.compare(o2.getPrice(), o1.getPrice());
+        return Double.compare(o2.getPrice(), o1.getPrice());
     }
 }
